+++ example/minimal/lib/main.dart
@@ -1,36 +1,9 @@
-import 'dart:io';
-
-import 'package:ansi_escape_codes/ansi_escape_codes.dart' as ansi;
 import 'package:flutter/material.dart';
 import 'package:scopo/scopo.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 void main() {
-  ScopeConfig.logger.level = ScopeLogLevel.info;
-
-  void setLogPrinter(
-    int level,
-    ansi.Color foreground, {
-    ansi.Color? background,
-  }) {
-    final printer = ansi.Printer(
-      ansiCodesEnabled: !Platform.isIOS,
-      defaultStyle: ansi.Style(
-        foreground: foreground,
-        background: background,
-      ),
-    );
-
-    ScopeConfig.logger[level].publisher = ScopeLogFormatter(
-      format: ScopeLogger.defaultFormat,
-      output: printer.print,
-    );
-  }
-
-  setLogPrinter(ScopeLogLevel.verbose, ansi.Color256.gray7);
-  setLogPrinter(ScopeLogLevel.debug, ansi.Color256.gray12);
-  setLogPrinter(ScopeLogLevel.info, ansi.Color256.rgb345);
-  setLogPrinter(ScopeLogLevel.error, ansi.Color256.rgb400);
+  ScopeConfig.observer = const ScopePrintObserver();
 
   runApp(const App(title: 'scopo minimal demo'));
 }